Allow lambda :bind property for body and heads
* hydra.el (hydra--head-property): Clean up doc.
(hydra--make-defun): Clean up doc.
(defhydra): Improve doc.
Both body and heads recognize :bind property in their plist.
It can be either nil or a lambda of `global-set-key' format.

Example:

    (defhydra hydra-goto (global-map "M-g"
                          :bind
                          (lambda (key cmd)
                            (bind-key key cmd)))
      ("g" goto-line "goto-line" :bind global-set-key)
      ("c" goto-char "goto-char"))

Here, `global-set-key' will be used to bind `goto-line' to "M-g g".
And `bind-key' will be used to bind `goto-char' to "M-g c".
Note that since `bind-key' is a macro, it was necessary to wrap it
in a lambda.

Since this commit, it's not possible to pass a lambda instead of
the whole BODY arg, as was advertised before. Just put it on
:bind now.