Make first line of hydra docstrings more informative.
Before this commit every head has the same first line for its docstring. The
distinguishing line is last. This change makes the distinguishing line appear
first in the docstring along with some minor wording changes.

For an application, see https://github.com/justbur/emacs-which-key/issues/185#issuecomment-361944776

* hydra.el (hydra--doc): Reword docstring.
(hydra--make-defun): Say which head is called in the first line.