Do not call fundamental-mode if lv buffer is already in fundamental-mode
Actually calling the fundamental-mode function can be a bit slow,
because it involves running change-major-mode-hook and also doing a
mode-line redisplay.  When the lv buffer is re-created over and
over, this can contribute to significant visual stuttering.

For example, in Spacemacs, pressing SPC w . (window transient state)
and then holding down } (enlarge) lags heavily before this change, but
is much smoother after.

Preserve the behavior of #364, however, to ensure that the lv buffer
is still in the correct major-mode.