Port defadvice to define-advice
As of Emacs 30.1, `defadvice` has been marked obselete, and can
lead to byte-compilation errors. Luckily, this is easy to fix;
all we need to do is to replace invocations of `defadvice` with
`define-advice`.

This also comes with some other minor benefits: we can remove the
surrounding `eval-after-load 'find-func` from our advice, as
`define-advice` can be used to advise functions that are not yet been
defined.